Refactor: [FE] selectbox 컴포넌트 분리
